# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "yachan-v2 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/boards/{slug}": {
      "get": {
        "tags": [
          "boards"
        ],
        "summary": "Get Board",
        "operationId": "get_board_api_boards__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/{board_slug}/posts/{post_number}": {
      "patch": {
        "tags": [
          "posts"
        ],
        "summary": "Edit Post",
        "operationId": "edit_post_api__board_slug__posts__post_number__patch",
        "parameters": [
          {
            "name": "board_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Slug"
            }
          },
          {
            "name": "post_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostEditCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mod/login": {
      "post": {
        "tags": [
          "mod"
        ],
        "summary": "Login",
        "operationId": "login_api_mod_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mod/boards": {
      "post": {
        "tags": [
          "mod"
        ],
        "summary": "Create Board",
        "operationId": "create_board_api_mod_boards_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mod/boards/order": {
      "put": {
        "tags": [
          "mod"
        ],
        "summary": "Reorder Boards",
        "operationId": "reorder_boards_api_mod_boards_order_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardReorder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardResponse"
                  },
                  "title": "Response Reorder Boards Api Mod Boards Order Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mod/boards/{board_slug}": {
      "patch": {
        "tags": [
          "mod"
        ],
        "summary": "Update Board",
        "operationId": "update_board_api_mod_boards__board_slug__patch",
        "parameters": [
          {
            "name": "board_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mod/{board_slug}/posts/{post_number}": {
      "delete": {
        "tags": [
          "mod"
        ],
        "summary": "Delete Post",
        "operationId": "delete_post_api_mod__board_slug__posts__post_number__delete",
        "parameters": [
          {
            "name": "board_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Slug"
            }
          },
          {
            "name": "post_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Number"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mod/{board_slug}/threads/{thread_id}/lock": {
      "post": {
        "tags": [
          "mod"
        ],
        "summary": "Lock Thread",
        "operationId": "lock_thread_api_mod__board_slug__threads__thread_id__lock_post",
        "parameters": [
          {
            "name": "board_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Slug"
            }
          },
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Thread Id"
            }
          },
          {
            "name": "locked",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Locked"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mod/{board_slug}/threads/{thread_id}/sticky": {
      "post": {
        "tags": [
          "mod"
        ],
        "summary": "Sticky Thread",
        "operationId": "sticky_thread_api_mod__board_slug__threads__thread_id__sticky_post",
        "parameters": [
          {
            "name": "board_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Slug"
            }
          },
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Thread Id"
            }
          },
          {
            "name": "sticky",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Sticky"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mod/{board_slug}/posts/{post_number}/ban": {
      "post": {
        "tags": [
          "mod"
        ],
        "summary": "Ban Poster",
        "operationId": "ban_poster_api_mod__board_slug__posts__post_number__ban_post",
        "parameters": [
          {
            "name": "board_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Board Slug"
            }
          },
          {
            "name": "post_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Number"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mod/reports": {
      "get": {
        "tags": [
          "mod"
        ],
        "summary": "List Reports",
        "operationId": "list_reports_api_mod_reports_get",
        "parameters": [
          {
            "name": "board_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Board Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportResponse"
                  },
                  "title": "Response List Reports Api Mod Reports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mod/reports/{report_id}/resolve": {
      "post": {
        "tags": [
          "mod"
        ],
        "summary": "Resolve Report",
        "operationId": "resolve_report_api_mod_reports__report_id__resolve_post",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BanCreate": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "board_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Board Id"
          }
        },
        "type": "object",
        "title": "BanCreate"
      },
      "BanResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ip_hash": {
            "type": "string",
            "title": "Ip Hash"
          },
          "board_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Board Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ip_hash",
          "board_id",
          "reason",
          "expires_at",
          "is_active",
          "created_at"
        ],
        "title": "BanResponse"
      },
      "BoardCreate": {
        "properties": {
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9_]{1,20}$",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "bump_limit": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Bump Limit",
            "default": 300
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title"
        ],
        "title": "BoardCreate"
      },
      "BoardReorder": {
        "properties": {
          "slugs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Slugs"
          }
        },
        "type": "object",
        "required": [
          "slugs"
        ],
        "title": "BoardReorder"
      },
      "BoardResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "bump_limit": {
            "type": "integer",
            "title": "Bump Limit"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "description",
          "bump_limit",
          "is_active",
          "position",
          "created_at"
        ],
        "title": "BoardResponse"
      },
      "BoardUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "bump_limit": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bump Limit"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "BoardUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModLogin": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "ModLogin"
      },
      "ModRole": {
        "type": "string",
        "enum": [
          "admin",
          "moderator"
        ],
        "title": "ModRole"
      },
      "PostEditCreate": {
        "properties": {
          "new_body": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "New Body"
          }
        },
        "type": "object",
        "title": "PostEditCreate"
      },
...
      "PostResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "post_number": {
            "type": "integer",
            "title": "Post Number"
          },
          "thread_id": {
            "type": "integer",
            "title": "Thread Id"
          },
          "board_id": {
            "type": "integer",
            "title": "Board Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tripcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tripcode"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "body_html": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Html"
          },
          "sage": {
            "type": "boolean",
            "title": "Sage"
          },
          "is_op": {
            "type": "boolean",
            "title": "Is Op"
          },
          "is_edited": {
            "type": "boolean",
            "title": "Is Edited"
          },
          "edited_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentResponse"
            },
            "type": "array",
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "post_number",
          "thread_id",
          "board_id",
          "name",
          "tripcode",
          "body",
          "body_html",
          "sage",
          "is_op",
          "is_edited",
          "edited_at",
          "created_at"
        ],
        "title": "PostResponse"
      },
...
      "ReportResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "post_id": {
            "type": "integer",
            "title": "Post Id"
          },
          "board_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Board Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "is_resolved": {
            "type": "boolean",
            "title": "Is Resolved"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "post_id",
          "board_id",
          "reason",
          "is_resolved",
          "created_at"
        ],
        "title": "ReportResponse"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "role": {
            "$ref": "#/components/schemas/ModRole"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "role"
        ],
        "title": "TokenResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}